Move all nested package files to base folder
