+++ onebuild/action_to_command_lookup.py
@@ -1,9 +1,9 @@
-from onebuild.actions.predefined_actions import PredefinedActions
-from onebuild.commands.help_command import HelpCommand
-from onebuild.commands.init_command import InitCommand
-from onebuild.commands.list_command import ListCommand
-from onebuild.commands.perform_command import PerformCommand
-from onebuild.commands.version_command import VersionCommand
+from onebuild.predefined_actions import PredefinedActions
+from onebuild.help_command import HelpCommand
+from onebuild.init_command import InitCommand
+from onebuild.list_command import ListCommand
+from onebuild.perform_command import PerformCommand
+from onebuild.version_command import VersionCommand
 
 
 class ActionToCommandLookup:
