+++ onebuild/list_command.py
@@ -1,4 +1,4 @@
-from onebuild.commands.command import Command
+from onebuild.command import Command
 from onebuild.config_parser import parse_project_config
 from onebuild.utils import config_string
